Minor fixes and deletes
