+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css'
 import Navbar from './components/Navbar'
-import Home2 from './pages/Home2'
+import Home from './pages/Home'
 import About from './pages/About'
 import Technologies from './pages/Technologies'
 import Projects from './pages/Projects'
@@ -12,7 +12,7 @@ function App() {
   return (
     <>
       <Navbar/>
-      <Home2/>
+      <Home/>
       <About/>
       <Technologies/>
       <Projects/>
